% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\docType{data}
\name{CGP.mat}
\alias{CGP.mat}
\title{Cancer Growth Pathway (CGP)}
\format{An object of class \code{matrix} with 932 rows and 932 columns.}
\usage{
data(CGP.mat)
}
\description{
A matrix contains the Cancer Growth Pathway (CGP) gene-gene interaction network information
}
\details{
The CGP was built using pathways empirically selected from KEGG pathway database based on our knowledge
firstly, and then refined by a small drug combination training dataset. In particular, we selected 12
pathways which were highly related to cell growth. Then we remove one of the 12 pathways each time and
merged the remaining to build a CGP with which we applied our approach to the external training dataset.
Based on the results, we screened out 8 pathways that contributed mostly to the performance and merged
them to build final CGP. The 8 KEGG pathways are: aminoacyl-tRNA biosynthesis, MAPK signaling pathway,
NF-kappa B signaling pathway, Cell Cycle, p53 signaling pathway, Apoptosis, TGF-beta signaling pathway,
Cancer pathway.
}
\keyword{datasets}
