% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\name{constGeneNet}
\alias{constGeneNet}
\title{Construct Gene Network Matrix}
\usage{
constGeneNet(geneNet)
}
\arguments{
\item{geneNet}{a data frame contains gene-gene interaction. See `geneNetLymph` for example.}
}
\value{
a matrix of gene connectivity matrix
}
\description{
This function is to convert a gene-gene interaction table to gene network matrix, which is an input of DIGREscore function.
}
\details{
Input gene-gene interaction table should have two columns with gene SYMBOL names. 
Each raw represents two connected genes. The interaction is regarded as undirected.
}
\examples{
geneNetLymph.mat <- constGeneNet(geneNet = geneNetLymph)

}
\author{
Jichen Yang, Sangin Lee, Minzhe Zhang(\email{zenroute.mzhang@gmail.com})
}
