% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\name{DIGREscore}
\alias{DIGREscore}
\title{Drug Induced Genomic Residual Effect (DIGRE) model}
\usage{
DIGREscore(geneExpDiff, doseRes, pathway = "KEGG", geneNet, fold = 0.6)
}
\arguments{
\item{geneExpDiff}{a matrix of drug treated gene expression data. Each column represent one drug,
each row represent one gene. The value represent the fold change of the expression of a particular
gene after drug treated compared to negative control.}

\item{doseRes}{a matrix contains drug dose response data. Each column represent one drug, two rows
of two different drug dose response curve.}

\item{pathway}{pathway information used in DIGRE model. User would specify either "KEGG" to use the KEGG
pathway information or "GeneNet" to use the gene network information.}

\item{geneNet}{optional parameter. If pathway parameter is "GeneNet", then specify this parameter to use 
your own gene network data. If pathway parameter is "KEGG", then do not set this parameter.}

\item{fold}{a value between 0 and 1. Gene expression fold change above this value would be considered
as upregulated, below the opposite number would be considered as downregulated, otherwise would be
considered as no effect. The default value is 0.6.}
}
\value{
a list contains two matrices. One gives the drug pair synergistic score and their rank, the other
contains the raw data to calculate the score.
}
\description{
This function estimate the compound pair synergy/antagonism .
It takes drug treated gene expression data and drug dose response curve data as input,
and caculate the synergistic score using Drug Induced Genomic Residual Effect (DIGRE) model.
}
\details{
This function takes drug treated gene expression data, dose-response curve data and pathway information as inputs,
and calculate pair synergistic score of all the possible combination of the compound you provided, and their rank
from the most synergistic to the most antagonist. Larger score indicates high possibility of the pair to have
synergistic effect, and vice versa. (Notice that this algorithm focus more on predicting the relative rank of your
compound pairs not the exact synergistic strength. If you want to do that, maybe you should involve positive control
in your experiment. And also the score calculated by two pathway information is not comparable.)
}
\author{
Jichen Yang, Sangin Lee, Minzhe Zhang(\email{Minzhe.Zhang@UTSouthwestern.edu})
}
\seealso{
Bansal M, Yang J, Karan C, et al. A community computational challenge to predict the activity of pairs of compounds[J].
Nature biotechnology, 2014, 32(12): 1213-1222.

Yang J, Tang H, Li Y, et al. DIGRE: Drug Induced Genomic Residual Effect Model for Successful Prediction of Multidrug
Effects[J]. CPT: pharmacometrics & systems pharmacology, 2015, 4(2): 91-97.
}

