% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\name{DIGREvis}
\alias{DIGREvis}
\title{Visualization of compound pairs synergistic score}
\usage{
DIGREvis(pred.pair, type)
}
\arguments{
\item{pred.pair}{data frame of score rank table generated by DIGREscore function.}

\item{type}{specify "heat" or "bar" to plot whether heatmap or barplot}
}
\value{
ggplot object of the heatmap or barplot of compound pairs synergistic score.
}
\description{
This function plot heatmap and barplot of compound pairs synergistic score.
}
\details{
Please directly apply this funtion to the result got from DIGREscore funtion.
}
\examples{
vis.heat <- DIGREvis(pred.pair = res.KEGG$scoreRank, type = "heat")
vis.bar <- DIGREvis(pred.pair = res.KEGG$scoreRank, type = "bar")
plot(vis.heat)
plot(vis.bar)

}
\author{
Jichen Yang, Sangin Lee, Minzhe Zhang(\email{zenroute.mzhang@gmail.com})
}
