% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\docType{data}
\name{doseRes.demo}
\alias{doseRes.demo}
\title{Dose response data}
\format{An object of class \code{data.frame} with 2 rows and 14 columns.}
\usage{
data(doseRes.demo)
}
\description{
Data frames of dose response example data for demo.
}
\details{
This data is from NCI-DREAM challenge competition for predicting drug pairs synergy.
}
\seealso{
\url{http://www.nature.com/nbt/journal/v32/n12/full/nbt.3052.html}
}
\keyword{datasets}

