% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\docType{data}
\name{geneExp.demo}
\alias{geneExp.demo}
\title{Drug treated gene expression data}
\format{An object of class \code{data.frame} with 48790 rows and 51 columns.}
\usage{
data(geneExp.demo)
}
\description{
Data frames of drug treated gene expression example data for demo.
}
\details{
This data is from NCI-DREAM challenge competition for predicting drug pairs synergy.
}
\seealso{
\url{http://www.nature.com/nbt/journal/v32/n12/full/nbt.3052.html}
}
\keyword{datasets}

