% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\docType{data}
\name{GP.mat}
\alias{GP.mat}
\title{Global Pathway (GP) information (KEGG)}
\format{An object of class \code{matrix} with 2709 rows and 2709 columns.}
\usage{
data(GP.mat)
}
\description{
A matrix contains the Global Pathway (GP) information built from KEGG pathway database.
}
\details{
To build GP information, we selected KEGG pathways belonging to genetic information processing, environmental
information processing, cellular processing, and cancer disease. From this set of selected pathways we removed
any pathway with fewer than 10 edges. Finally we merged the remaining 32 KEGG pathways into a global pathway (GP)
which included 11642 interactions among 2322 genes.
}
\keyword{datasets}

