% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\docType{data}
\name{mCGP.mat}
\alias{mCGP.mat}
\title{Cancer Growth Pathway (CGP) information (Gene Network)}
\format{An object of class \code{matrix} with 875 rows and 875 columns.}
\usage{
data(mCGP.mat)
}
\description{
A matrix contains the Cancer Growth Pathway (CGP) information built from lymphoma gene network.
}
\details{
We used GSE10846 to build the lymphoma gene network.
}
\seealso{
\link{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE10846}
}
\keyword{datasets}

