% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\docType{data}
\name{mGP.mat}
\alias{mGP.mat}
\title{Global Pathway (GP) information (Gene Network)}
\format{An object of class \code{matrix} with 2559 rows and 2559 columns.}
\usage{
data(mGP.mat)
}
\description{
A matrix contains the Global Pathway (GP) information built from lymphoma gene network.
}
\details{
We used GSE10846 to build the lymphoma gene network.
}
\seealso{
\link{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE10846}
}
\keyword{datasets}

