% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\name{profileGeneExp}
\alias{profileGeneExp}
\title{Profiling Drug Treated Gene Expression Data}
\usage{
profileGeneExp(geneExp)
}
\arguments{
\item{geneExp}{a data frame contains the drug treated gene expression data with each column representing 
one drug, and each row representing one gene. See `geneExp.demo` for example.}
}
\value{
a matrix of processed gene expression data.
}
\description{
This function profile the drug treated gene expression data to prepare the input for DIGREscore function.
}
\details{
Gene expression is measured after cell is treated by a single compound (or negative control).
Raw data (micro array or RNA-Seq) should already be log transformed to have proper scale. 
This function will average duplicated data with same drug name, and collapse multiple probes to gene level.
}
\examples{
geneExpDiff <- profileGeneExp(geneExp = geneExp.demo)

}
\author{
Jichen Yang, Sangin Lee, Minzhe Zhang(\email{zenroute.mzhang@gmail.com})
}
