% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\name{readDoseRes.csv}
\alias{readDoseRes.csv}
\title{Read Drug Dose Response Data}
\usage{
readDoseRes.csv(file)
}
\arguments{
\item{file}{a file contains the drug dose response data (.csv file accepted)}
}
\value{
a matrix of drug dose response data with desired format that DIGRE can understand.
}
\description{
This function profile the drug treated gene expression data to prepare the input for DIGREscore function.
}
\author{
Jichen Yang, Sangin Lee, Minzhe Zhang(\email{Minzhe.Zhang@UTSouthwestern.edu})
}

