% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\name{constGeneNet}
\alias{constGeneNet}
\title{Constructing Gene Network Matrix for (DIGRE) model}
\usage{
constGeneNet(geneNet)
}
\arguments{
\item{geneNet}{a data frame contains gene-gene interaction inforamtion.}
}
\value{
a matrix of gene connectivity matrix
}
\description{
This function construct gene network matrix from gene interaction to prepare the pathway information for DIGREscore function.
}
\details{
Input data should have two columns of gene SYMBOL names. Each raw represents two connected nodes.
}
\author{
Jichen Yang, Sangin Lee, Minzhe Zhang(\email{Minzhe.Zhang@UTSouthwestern.edu})
}

