% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\name{DIGREvis}
\alias{DIGREvis}
\title{Visualization of compound pairs synergistic score}
\usage{
DIGREvis(pred.pair, type)
}
\arguments{
\item{pred.pair}{data frame of score rank table generated by DIGREscore function.}

\item{type}{specify "heat" or "bar" to plot whether heatmap or barplot}
}
\value{
ggplot object of the heatmap or barplot of compound pairs synergistic score.
}
\description{
This function plot heatmap and barplot of compound pairs synergistic score.
}
\details{
Please direct apply this funtion to the result got from DIGREscore funtion.
}
\author{
Jichen Yang, Sangin Lee, Minzhe Zhang(\email{Minzhe.Zhang@UTSouthwestern.edu})
}

