% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\docType{data}
\name{doseRes.demo}
\alias{doseRes.demo}
\title{Dose response data}
\format{An object of class \code{data.frame} with 2 rows and 14 columns.}
\usage{
data(doseRes.demo)
}
\description{
Examplary drug dose response data of 14 drugs in NCI-DREAM challenge.
}
\details{
This demo data contains dose response data of 14 drugs from NCI-DREAM challenge. 
It contain the cell viability reduction when treated with drug in two different dose. 
One dose is IC20 of the drug, therefore the cell viability reduction is always 0.2 for all drugs. 
The other dose is double dose of IC20, this value is infered from the dose response curve of each drug.
}
\seealso{
\url{http://www.nature.com/nbt/journal/v32/n12/full/nbt.3052.html}
}
\keyword{datasets}
