% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\docType{data}
\name{geneExp.demo}
\alias{geneExp.demo}
\title{Drug treated gene expression data}
\format{An object of class \code{data.frame} with 48704 rows and 51 columns.}
\usage{
data(geneExp.demo)
}
\description{
Examplary gene expression data of DLBCL cell after treated with 14 different drugs.
}
\details{
This data is from NCI-DREAM challenge competition for predicting drug pairs synergy. 
OCI-LY3 human diffuse large B-cell lymphoma (DLBCL) cell line was treated by 14 different 
drugs in its dose of IC20. 24 hours after Perturbation, gene expression level was measured.
}
\seealso{
\url{http://www.nature.com/nbt/journal/v32/n12/full/nbt.3052.html}
}
\keyword{datasets}
