% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\docType{data}
\name{geneNetLymph}
\alias{geneNetLymph}
\title{Lymphoma Specific Gene Network}
\format{An object of class \code{data.frame} with 29055 rows and 2 columns.}
\usage{
data(geneNetLymph)
}
\description{
Gene-gene interaction information refined from lymphoma patients.
}
\details{
We use GSE10846 to construct this network.
}
\seealso{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE10846}
}
\keyword{datasets}

