% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIGREsyn.R
\name{profileGeneExp}
\alias{profileGeneExp}
\title{Profiling Drug Treated Gene Expression Data for (DIGRE) model}
\usage{
profileGeneExp(file)
}
\arguments{
\item{file}{a file contains the drug treated gene expression data (.csv file accepted)}
}
\value{
a matrix of processed gene expression data.
}
\description{
This function profile the drug treated gene expression data to prepare the input for DIGREscore function.
}
\details{
Input file should be gene expression profiles of single compound-treated and negative control-treated cell line sample.
Data needs to be log2 transformed. This function will average duplicated data with same drug name. And collapse multiple
probes data to gene level.
}
\author{
Jichen Yang, Sangin Lee, Minzhe Zhang(\email{Minzhe.Zhang@UTSouthwestern.edu})
}

